+++ src/command_modules/azure-cli-eventgrid/azure/cli/command_modules/eventgrid/_params.py
@@ -16,8 +16,6 @@ from azure.cli.core.commands.parameters import (
     name_type
 )
 
-from azure.cli.command_modules.eventgrid.sdk.models import EndpointType
-
 included_event_types_type = CliArgumentType(
     help="A space separated list of event types. To subscribe to all event types, the string \"All\" should be specified.",
     nargs='+'
@@ -32,7 +30,7 @@ register_cli_argument('eventgrid', 'resource_group_name', resource_group_name_ty
 register_cli_argument('eventgrid', 'location', location_type)
 register_cli_argument('eventgrid', 'tags', tags_type, help="Space separated tags in 'key[=value]' format.")
 
-register_cli_argument('eventgrid', 'endpoint', help="Endpoint where EventGrid should deliver events matching this event subscription.")
+register_cli_argument('eventgrid', 'endpoint', help="Endpoint where EventGrid should deliver events matching this event subscription. For webhook endpoint type, this should be the corresponding webhook URL. For eventhub endpoint type, this should be the Azure ResourceID of the event hub.")
 register_cli_argument('eventgrid', 'event_subscription_name', help="Name of the event subscription.")
 register_cli_argument('eventgrid', 'subject_begins_with', help="An optional string to filter events for an event subscription based on a prefix. Wildcard characters are not supported.")
 register_cli_argument('eventgrid', 'subject_ends_with', help="An optional string to filter events for an event subscription based on a suffix. Wildcard characters are not supported.")
@@ -49,7 +47,7 @@ register_cli_argument('eventgrid', 'is_subject_case_sensitive', options_list=['-
 
 register_cli_argument('eventgrid', 'included_event_types', included_event_types_type)
 register_cli_argument('eventgrid', 'labels', labels_type)
-register_cli_argument('eventgrid', 'endpoint_type', **enum_choice_list(EndpointType))
+register_cli_argument('eventgrid', 'endpoint_type', **enum_choice_list(['webhook', 'eventhub']))
 
 register_cli_argument('eventgrid', 'provider_namespace', help="Namespace of the provider owning the resource.")
 register_cli_argument('eventgrid', 'resource_type', help="Type of the resource.")