+++ src/command_modules/azure-cli-resource/azure/cli/command_modules/resource/_help.py
@@ -38,6 +38,11 @@ helps['managedapp definition create'] = """
             az managedapp definition create -g MyResourceGroup -n MyManagedAppDef -l eastus --display-name "MyManagedAppDef" \\
                 --description "My Managed App Def description" -a "myPrincipalId:myRoleId" --lock-level None \\
                 --package-file-uri "https://path/to/myPackage.zip"
+        - name: Create a managed application defintion with inline values for createUiDefinition and mainTemplate.
+          text: >
+            az managedapp definition create -g MyResourceGroup -n MyManagedAppDef -l eastus --display-name "MyManagedAppDef" \\
+                --description "My Managed App Def description" -a "myPrincipalId:myRoleId" --lock-level None \\
+                --create-ui-definition @myCreateUiDef.json --main-template @myMainTemplate.json
 """
 helps['managedapp definition delete'] = """
     type: command
@@ -428,6 +433,15 @@ helps['provider'] = """
     type: group
     short-summary: Manage resource providers.
 """
+
+helps['provider list'] = """
+    type: command
+    examples:
+        - name: Display all resource types for the network resource provider.
+          text: >
+            az provider list --query [?namespace=='Microsoft.Network'].resourceTypes[].resourceType
+"""
+
 helps['provider register'] = """
     type: command
     short-summary: Register a provider.