+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/commands.py
@@ -21,7 +21,7 @@ from azure.cli.core.profiles import supported_api_version, ResourceType
 # pylint: disable=line-too-long
 
 custom_path = 'azure.cli.command_modules.vm.custom#{}'
-mgmt_path = 'azure.mgmt.compute.compute.operations.{}#{}.{}'
+mgmt_path = 'azure.mgmt.compute.operations.{}#{}.{}'
 
 
 # VM
@@ -41,13 +41,16 @@ def transform_ip_addresses(result):
     return transformed
 
 
-def transform_vm(result):
-    return OrderedDict([('name', result['name']),
-                        ('resourceGroup', result['resourceGroup']),
-                        ('powerState', result.get('powerState')),
-                        ('publicIps', result.get('publicIps')),
-                        ('fqdns', result.get('fqdns')),
-                        ('location', result['location'])])
+def transform_vm(vm):
+    result = OrderedDict([('name', vm['name']),
+                          ('resourceGroup', vm['resourceGroup']),
+                          ('powerState', vm.get('powerState')),
+                          ('publicIps', vm.get('publicIps')),
+                          ('fqdns', vm.get('fqdns')),
+                          ('location', vm['location'])])
+    if 'zones' in vm:
+        result['zones'] = ','.join(vm['zones']) if vm['zones'] else ''
+    return result
 
 
 def transform_vm_create_output(result):
@@ -63,6 +66,8 @@ def transform_vm_create_output(result):
                               ('location', result.location)])
         if getattr(result, 'identity', None):
             output['identity'] = result.identity
+        if hasattr(result, 'zones'):  # output 'zones' column even the property value is None
+            output['zones'] = result.zones[0] if result.zones else ''
         return output
     except AttributeError:
         from msrest.pipeline import ClientRawResponse
@@ -132,14 +137,23 @@ cli_generic_update_command(__name__, 'vm update',
                            cf_vm,
                            no_wait_param='raw')
 cli_generic_wait_command(__name__, 'vm wait', 'azure.cli.command_modules.vm.custom#get_instance_view')
+if supported_api_version(ResourceType.MGMT_COMPUTE, min_api='2017-03-30'):
+    cli_command(__name__, 'vm perform-maintenance', mgmt_path.format(op_var, op_class, 'perform_maintenance'), cf_vm)
+
 
 if supported_api_version(ResourceType.MGMT_COMPUTE, min_api='2016-04-30-preview'):
     cli_command(__name__, 'vm convert', mgmt_path.format(op_var, op_class, 'convert_to_managed_disks'), cf_vm)
 
-    # VM encryption
-    cli_command(__name__, 'vm encryption enable', 'azure.cli.command_modules.vm.disk_encryption#enable')
-    cli_command(__name__, 'vm encryption disable', 'azure.cli.command_modules.vm.disk_encryption#disable')
-    cli_command(__name__, 'vm encryption show', 'azure.cli.command_modules.vm.disk_encryption#show', exception_handler=empty_on_404)
+# VM encryption
+cli_command(__name__, 'vm encryption enable', 'azure.cli.command_modules.vm.disk_encryption#encrypt_vm')
+cli_command(__name__, 'vm encryption disable', 'azure.cli.command_modules.vm.disk_encryption#decrypt_vm')
+cli_command(__name__, 'vm encryption show', 'azure.cli.command_modules.vm.disk_encryption#show_vm_encryption_status', exception_handler=empty_on_404)
+
+# VMSS encryption
+if supported_api_version(ResourceType.MGMT_COMPUTE, min_api='2017-03-30'):
+    cli_command(__name__, 'vmss encryption enable', 'azure.cli.command_modules.vm.disk_encryption#encrypt_vmss')
+    cli_command(__name__, 'vmss encryption disable', 'azure.cli.command_modules.vm.disk_encryption#decrypt_vmss')
+    cli_command(__name__, 'vmss encryption show', 'azure.cli.command_modules.vm.disk_encryption#show_vmss_encryption_status', exception_handler=empty_on_404)
 
 # VM NIC
 cli_command(__name__, 'vm nic add', custom_path.format('vm_add_nics'))
@@ -250,7 +264,8 @@ cli_command(__name__, 'vm list-usage', mgmt_path.format('usage_operations', 'Usa
             table_transformer='[].{Name:localName, CurrentValue:currentValue, Limit:limit}')
 
 # VMSS
-vmss_show_table_transform = '{Name:name, ResourceGroup:resourceGroup, Location:location, Capacity:sku.capacity, Overprovision:overprovision, upgradePolicy:upgradePolicy.mode}'
+vmss_show_table_transform = '{Name:name, ResourceGroup:resourceGroup, Location:location, $zone$Capacity:sku.capacity, Overprovision:overprovision, UpgradePolicy:upgradePolicy.mode}'
+vmss_show_table_transform = vmss_show_table_transform.replace('$zone$', 'Zones: (!zones && \' \') || join(` `, zones), ' if supported_api_version(ResourceType.MGMT_COMPUTE, min_api='2017-03-30') else ' ')
 cli_command(__name__, 'vmss delete', mgmt_path.format('virtual_machine_scale_sets_operations', 'VirtualMachineScaleSetsOperations', 'delete'), cf_vmss, no_wait_param='raw')
 cli_command(__name__, 'vmss list-skus', mgmt_path.format('virtual_machine_scale_sets_operations', 'VirtualMachineScaleSetsOperations', 'list_skus'), cf_vmss)
 
@@ -279,19 +294,21 @@ cli_command(__name__, 'vm list-sizes', mgmt_path.format('virtual_machine_sizes_o
 
 if supported_api_version(ResourceType.MGMT_COMPUTE, min_api='2017-03-30'):
     # VM Disk
+    disk_show_table_transform = '{Name:name, ResourceGroup:resourceGroup, Location:location, $zone$Sku:sku.name, OsType:osType, SizeGb:diskSizeGb, ProvisioningState:provisioningState}'
+    disk_show_table_transform = disk_show_table_transform.replace('$zone$', 'Zones: (!zones && \' \') || join(` `, zones), ' if supported_api_version(ResourceType.MGMT_COMPUTE, min_api='2017-03-30') else ' ')
     op_var = 'disks_operations'
     op_class = 'DisksOperations'
-    cli_command(__name__, 'disk create', custom_path.format('create_managed_disk'), no_wait_param='no_wait')
-    cli_command(__name__, 'disk list', custom_path.format('list_managed_disks'))
-    cli_command(__name__, 'disk show', mgmt_path.format(op_var, op_class, 'get'), cf_disks, exception_handler=empty_on_404)
+    cli_command(__name__, 'disk create', custom_path.format('create_managed_disk'), no_wait_param='no_wait', table_transformer=disk_show_table_transform)
+    cli_command(__name__, 'disk list', custom_path.format('list_managed_disks'), table_transformer='[].' + disk_show_table_transform)
+    cli_command(__name__, 'disk show', mgmt_path.format(op_var, op_class, 'get'), cf_disks, exception_handler=empty_on_404, table_transformer=disk_show_table_transform)
     cli_command(__name__, 'disk delete', mgmt_path.format(op_var, op_class, 'delete'), cf_disks, no_wait_param='raw', confirmation=True)
     cli_command(__name__, 'disk grant-access', custom_path.format('grant_disk_access'))
     cli_command(__name__, 'disk revoke-access', mgmt_path.format(op_var, op_class, 'revoke_access'), cf_disks)
-    cli_generic_update_command(__name__, 'disk update', 'azure.mgmt.compute.compute.operations.{}#{}.get'.format(op_var, op_class),
-                               'azure.mgmt.compute.compute.operations.{}#{}.create_or_update'.format(op_var, op_class),
+    cli_generic_update_command(__name__, 'disk update', 'azure.mgmt.compute.operations.{}#{}.get'.format(op_var, op_class),
+                               'azure.mgmt.compute.operations.{}#{}.create_or_update'.format(op_var, op_class),
                                custom_function_op=custom_path.format('update_managed_disk'),
                                setter_arg_name='disk', factory=cf_disks, no_wait_param='raw')
-    cli_generic_wait_command(__name__, 'disk wait', 'azure.mgmt.compute.compute.operations.{}#{}.get'.format(op_var, op_class), cf_disks)
+    cli_generic_wait_command(__name__, 'disk wait', 'azure.mgmt.compute.operations.{}#{}.get'.format(op_var, op_class), cf_disks)
 
     op_var = 'snapshots_operations'
     op_class = 'SnapshotsOperations'
@@ -301,8 +318,8 @@ if supported_api_version(ResourceType.MGMT_COMPUTE, min_api='2017-03-30'):
     cli_command(__name__, 'snapshot delete', mgmt_path.format(op_var, op_class, 'delete'), cf_snapshots)
     cli_command(__name__, 'snapshot grant-access', custom_path.format('grant_snapshot_access'))
     cli_command(__name__, 'snapshot revoke-access', mgmt_path.format(op_var, op_class, 'revoke_access'), cf_snapshots)
-    cli_generic_update_command(__name__, 'snapshot update', 'azure.mgmt.compute.compute.operations.{}#{}.get'.format(op_var, op_class),
-                               'azure.mgmt.compute.compute.operations.{}#{}.create_or_update'.format(op_var, op_class),
+    cli_generic_update_command(__name__, 'snapshot update', 'azure.mgmt.compute.operations.{}#{}.get'.format(op_var, op_class),
+                               'azure.mgmt.compute.operations.{}#{}.create_or_update'.format(op_var, op_class),
                                custom_function_op=custom_path.format('update_snapshot'),
                                setter_arg_name='snapshot', factory=cf_snapshots)