+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/custom.py
@@ -299,7 +299,7 @@ def create_managed_disk(resource_group_name, disk_name, location=None,
                         source=None,  # pylint: disable=unused-argument
                         # below are generated internally from 'source'
                         source_blob_uri=None, source_disk=None, source_snapshot=None,
-                        source_storage_account_id=None, no_wait=False, tags=None):
+                        source_storage_account_id=None, no_wait=False, tags=None, zone=None):
     Disk, CreationData, DiskCreateOption = get_sdk(ResourceType.MGMT_COMPUTE, 'Disk', 'CreationData',
                                                    'DiskCreateOption', mod='models')
 
@@ -319,6 +319,9 @@ def create_managed_disk(resource_group_name, disk_name, location=None,
     if size_gb is None and option == DiskCreateOption.empty:
         raise CLIError('usage error: --size-gb required to create an empty disk')
     disk = Disk(location, creation_data, (tags or {}), _get_sku_object(sku), disk_size_gb=size_gb)
+    if zone:
+        disk.zones = zone
+
     client = _compute_client_factory()
     return client.disks.create_or_update(resource_group_name, disk_name, disk, raw=no_wait)
 
@@ -477,7 +480,7 @@ def grant_snapshot_access(resource_group_name, snapshot_name, duration_in_second
 
 
 def _grant_access(resource_group_name, name, duration_in_seconds, is_disk):
-    from azure.mgmt.compute.models import AccessLevel
+    AccessLevel = get_sdk(ResourceType.MGMT_COMPUTE, 'AccessLevel', mod='models')
     client = _compute_client_factory()
     op = client.disks if is_disk else client.snapshots
     return op.grant_access(resource_group_name, name, AccessLevel.read, duration_in_seconds)
@@ -1555,10 +1558,10 @@ def convert_av_set_to_managed_disk(resource_group_name, availability_set_name):
 
 
 # pylint: disable=too-many-locals, unused-argument, too-many-statements, too-many-branches
-def create_vm(vm_name, resource_group_name, image=None, size='Standard_DS1_v2', location=None, tags=None, no_wait=False,
-              authentication_type=None, admin_password=None, admin_username=DefaultStr(getpass.getuser()),
-              ssh_dest_key_path=None, ssh_key_value=None, generate_ssh_keys=False,
-              availability_set=None, nics=None, nsg=None, nsg_rule=None,
+def create_vm(vm_name, resource_group_name, image=None, size='Standard_DS1_v2', location=None, tags=None,
+              no_wait=False, authentication_type=None, admin_password=None,
+              admin_username=DefaultStr(getpass.getuser()), ssh_dest_key_path=None, ssh_key_value=None,
+              generate_ssh_keys=False, availability_set=None, nics=None, nsg=None, nsg_rule=None,
               private_ip_address=None, public_ip_address=None, public_ip_address_allocation='dynamic',
               public_ip_address_dns_name=None, os_disk_name=None, os_type=None, storage_account=None, os_caching=None,
               data_caching=None, storage_container_name=None, storage_sku=None, use_unmanaged_disk=False,
@@ -1568,7 +1571,8 @@ def create_vm(vm_name, resource_group_name, image=None, size='Standard_DS1_v2', 
               storage_account_type=None, vnet_type=None, nsg_type=None, public_ip_type=None, nic_type=None,
               validate=False, custom_data=None, secrets=None, plan_name=None, plan_product=None, plan_publisher=None,
               license_type=None, assign_identity=False, identity_scope=None,
-              identity_role=DefaultStr('Contributor'), identity_role_id=None):
+              identity_role=DefaultStr('Contributor'), identity_role_id=None, application_security_groups=None,
+              zone=None):
     from azure.cli.core.commands.client_factory import get_subscription_id
     from azure.cli.core.util import random_string, hash_string
     from azure.cli.command_modules.vm._template_builder import (ArmTemplateBuilder, build_vm_resource,
@@ -1630,7 +1634,7 @@ def create_vm(vm_name, resource_group_name, image=None, size='Standard_DS1_v2', 
                                                                   tags,
                                                                   public_ip_address_allocation,
                                                                   public_ip_address_dns_name,
-                                                                  None))
+                                                                  None, zone))
 
         subnet_id = subnet if is_valid_resource_id(subnet) else \
             '{}/virtualNetworks/{}/subnets/{}'.format(network_id_template, vnet_name, subnet)
@@ -1650,15 +1654,15 @@ def create_vm(vm_name, resource_group_name, image=None, size='Standard_DS1_v2', 
         ]
         nic_resource = build_nic_resource(
             nic_name, location, tags, vm_name, subnet_id, private_ip_address, nsg_id,
-            public_ip_address_id)
+            public_ip_address_id, application_security_groups)
         nic_resource['dependsOn'] = nic_dependencies
         master_template.add_resource(nic_resource)
     else:
         # Using an existing NIC
-        invalid_parameters = [nsg, public_ip_address, subnet, vnet_name]
+        invalid_parameters = [nsg, public_ip_address, subnet, vnet_name, application_security_groups]
         if any(invalid_parameters):
             raise CLIError('When specifying an existing NIC, do not specify NSG, '
-                           'public IP, VNet or subnet.')
+                           'public IP, ASGs, VNet or subnet.')
 
     os_vhd_uri = None
     if storage_profile in [StorageProfile.SACustomImage, StorageProfile.SAPirImage]:
@@ -1683,7 +1687,7 @@ def create_vm(vm_name, resource_group_name, image=None, size='Standard_DS1_v2', 
         admin_password, ssh_key_value, ssh_dest_key_path, image, os_disk_name,
         os_type, os_caching, data_caching, storage_sku, os_publisher, os_offer, os_sku, os_version,
         os_vhd_uri, attach_os_disk, attach_data_disks, data_disk_sizes_gb, image_data_disks, custom_data, secrets,
-        license_type)
+        license_type, zone)
     vm_resource['dependsOn'] = vm_dependencies
 
     if plan_name:
@@ -1769,7 +1773,7 @@ def create_vmss(vmss_name, resource_group_name, image,
                 single_placement_group=None, custom_data=None, secrets=None,
                 plan_name=None, plan_product=None, plan_publisher=None,
                 assign_identity=False, identity_scope=None, identity_role=DefaultStr('Contributor'),
-                identity_role_id=None):
+                identity_role_id=None, zones=None):
     from azure.cli.core.commands.client_factory import get_subscription_id
     from azure.cli.core.util import random_string, hash_string
     from azure.cli.command_modules.vm._template_builder import (ArmTemplateBuilder, StorageProfile, build_vmss_resource,
@@ -1847,6 +1851,8 @@ def create_vmss(vmss_name, resource_group_name, image,
         vmss_dependencies.append('Microsoft.Network/loadBalancers/{}'.format(load_balancer))
 
         lb_dependencies = []
+        if vnet_type == 'new':
+            lb_dependencies.append('Microsoft.Network/virtualNetworks/{}'.format(vnet_name))
         if public_ip_type == 'new':
             public_ip_address = public_ip_address or '{}PublicIP'.format(load_balancer)
             lb_dependencies.append(
@@ -1854,7 +1860,7 @@ def create_vmss(vmss_name, resource_group_name, image,
             master_template.add_resource(build_public_ip_resource(
                 public_ip_address, location, tags,
                 _get_public_ip_address_allocation(public_ip_address_allocation, load_balancer_sku),
-                public_ip_address_dns_name, load_balancer_sku))
+                public_ip_address_dns_name, load_balancer_sku, zones))
             public_ip_address_id = '{}/publicIPAddresses/{}'.format(network_id_template,
                                                                     public_ip_address)
 
@@ -1876,6 +1882,8 @@ def create_vmss(vmss_name, resource_group_name, image,
         vmss_dependencies.append('Microsoft.Network/applicationGateways/{}'.format(app_gateway))
 
         ag_dependencies = []
+        if vnet_type == 'new':
+            ag_dependencies.append('Microsoft.Network/virtualNetworks/{}'.format(vnet_name))
         if public_ip_type == 'new':
             public_ip_address = public_ip_address or '{}PublicIP'.format(app_gateway)
             ag_dependencies.append(
@@ -1883,7 +1891,7 @@ def create_vmss(vmss_name, resource_group_name, image,
             master_template.add_resource(build_public_ip_resource(
                 public_ip_address, location, tags,
                 _get_public_ip_address_allocation(public_ip_address_allocation, None), public_ip_address_dns_name,
-                None))
+                None, zones))
             public_ip_address_id = '{}/publicIPAddresses/{}'.format(network_id_template,
                                                                     public_ip_address)
 
@@ -1959,7 +1967,7 @@ def create_vmss(vmss_name, resource_group_name, image,
                                         os_publisher, os_offer, os_sku, os_version,
                                         backend_address_pool_id, inbound_nat_pool_id,
                                         single_placement_group=single_placement_group,
-                                        custom_data=custom_data, secrets=secrets)
+                                        custom_data=custom_data, secrets=secrets, zones=zones)
     vmss_resource['dependsOn'] = vmss_dependencies
 
     if plan_name: