+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+0.1.17 (2017-09-22)
++++++++++++++++++++
+* webapp: able to update and show authentication settings using "az webapp auth update/show"
+
 0.1.16 (2017-09-11)
 +++++++++++++++++++
 * webapp: able to create a webapp in a resource group other than the service plan's